+++ app/controllers/bookingsync_portal/admin/base_controller.rb
@@ -5,10 +5,15 @@ module BookingsyncPortal
     class BaseController < ApplicationController
       layout 'bookingsync_portal/admin'
       respond_to :html
-
       include BookingsyncApplication::Admin::CommonBaseController
 
+      before_action :enforce_remote_account!
+
       private
+
+      def enforce_remote_account!
+        redirect_to new_admin_remote_account_path if current_account.remote_accounts.empty?
+      end
 
       def messagebus_channel
         "/account-#{current_account.id}"