+++ lib/bookingsync_portal.rb
@@ -1,5 +1,6 @@
 require 'bookingsync_application'
 require 'bookingsync_portal/engine'
+require 'bookingsync_portal/mash_serializer'
 require 'message_bus'
 
 # FIXME requires below should get removed when ember frontend is added