+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Added redirect to remote_accounts#create when no remote account is present.
+
 # 0.0.1
 
 * First public versioned release.